API simplification as suggested in comments
