+++ py/sight/worker/text_proto_generator.py
