+++ py/worker/attribute_configs/fvs_attribute_config.py
@@ -1,6 +1,8 @@
 from typing import Any
 
-from sight.proto import sight_pb2
+from sight.worker.worker_helper import create_attr_props
+
+
 
 FVS_PARAMS = {
     "fire-SIMFIRE_{index}-6_stand_area_burned": None,
@@ -11,25 +13,6 @@ FVS_PARAMS = {
 }
 
 MAX_FVS_CYCLE = 40
-
-
-def create_attr_props(
-    config_dict: dict[str, Any],
-) -> dict[str, sight_pb2.DecisionConfigurationStart.AttrProps]:
-  """Create AttrProps for each key in the config_dict.
-
-  Args:
-    config_dict (dict[str, Any]): Config dict to be converted to AttrProps.
-
-  Returns:
-    dict[str, sight_pb2.DecisionConfigurationStart.AttrProps]:
-    AttrProps for each key in the config_dict.
-  """
-  return {
-      key: (sight_pb2.DecisionConfigurationStart.AttrProps() if value
-            is not None else sight_pb2.DecisionConfigurationStart.AttrProps()
-           ) for key, value in config_dict.items()
-  }
 
 
 def expand_params_for_cycles(fvs_params: dict[str, Any]) -> dict[str, Any]:
@@ -60,36 +43,8 @@ def get_action_attrs():
   return create_attr_props(action_config)
 
 
-# action_attrs = {
-#     "a1":
-#     sight_pb2.DecisionConfigurationStart.AttrProps(
-#         min_value=0,
-#         max_value=1,
-#     ),
-#     "a2":
-#     sight_pb2.DecisionConfigurationStart.AttrProps(
-#         min_value=0,
-#         max_value=1,
-#     ),
-#     "a3":
-#     sight_pb2.DecisionConfigurationStart.AttrProps(
-#         min_value=0,
-#         max_value=1,
-#     ),
-# }
-
-
 def get_outcome_attrs():
   """Returns the outcome attributes for the FVS outcome.
   """
   outcome_config = {'time_series': None}
   return create_attr_props(outcome_config)
-
-
-# outcome_attrs = {
-#     "time_series":
-#     sight_pb2.DecisionConfigurationStart.AttrProps(
-#         min_value=0,
-#         max_value=1,
-#     ),
-# }
